Pitfalls: Don't use 'm' as a template variable
Using 'm' as a template variable blocks model usage for the entire scope of
that variable.  If model usage is attempted it leads to confusing template
errors.

The purpose of this change is to reinforce the Best Practices and Pitfalls
guide in the Cookbook by removing bad practices from the code base.