core: Fix use of hostname in startup / management scripts
HOSTNAME is no longer used directly to address the node. Instead, an
internal NODEHOST variable is used which is equal to the system
hostname ($HOSTNAME).

However, in the case of running with -sname (the default), only the
first part of $HOSTNAME (everything before the first dot) is used as
the node host.

Additionally, added some checks to zotonic start/stop/debug to print a
human-readable explanation when attempting to start the same Zotonic
node twice or trying to stop Zotonic when it is not running.
(cherry picked from commit c34247eab5cecca128bd75f2d09d7b5edcbf950f)