doc: stylistic tweak, but I'm not sure this paragraph makes sense.
The original wording suggests that a module purely consisting of these
metadata properties is a common and useful thing. I don't know much
about Zotonic yet, but apart from a module which contains stand alone
templates I can't imagine how such a module is used.

My guess is that the intention is to say that Zotonic comes with many
useful functions which can be used from a custom module without the
need for writing lots of code ... but I'm really not sure ...