core: change stemming of the full text indexes.
This fixes a problem where the stemming of the full text index didn't match the stemming of the search query.
The stemmer is selected using the language code from either the i18n.language_stemmer or i18n.language configuration.
Call z_pivot_rsc:stemmer_language(Context) to see which PostgreSQL stemmer will be used.
You have to ensure that the selected stemmer is available in PostgresSQL.
With psql you can use the command '\dFd' to see the list of available stemmers.

Fixes #553