Anonymous user will be able to edit their temporary resource until their session is over.
Since there is no way to check for creator id's for anonymous users we need
another mechanism to prevent users from losing control of their temporary
resource between creating and editing it.

This is done by linking the temporary resource to a session_id and declare
whoever has that session Id as a temporary owner of that resource. When the
session ends the temporary resource cleanup task will remove the link between
session and resource if it is not up for deletion. This way ACL rules can be
configured for that time frame.