When duplicate reset custom_slug boolean
Duplicate page does not work when the page has a title_slug, even though it is added as undefined to the SafeDupProps

Problem 1. the prop title_slug is added as undefined, then in props filter if the title_slug is empty it is added as a empty slug and title_slug. slug is not allowed as a empty value. Resulting in an empty slug and a database error.

Problem 2. checks for slugs are done by checking custom_slug boolean, this is currently not reset to false. For the slug the value is then not reset to default using the title of the resource for the slug. Resulting in an empty slug and a database error.