changing sign in to login
changing sign in to login
It's really hard for a non native english speaker to distinct sign up/sign in/sign out. In zotonic it was used inconsistantly, so now it's always:
Create account (not sign up) and  log in/log out (if it's a verb) and if it's a noun (the button): Login.
Signup is only used if it's for a calendar event (e.g in the LearningStone calendar application)