Add IF NOT EXISTS condition to key_changes_v1_0's index install
Problem: In 'z_install_update:upgrade/3' the 'fki_rsc_category_id'
index is installed twice when upgrading from version 0.82 or
lower, because of 'check_category_id_key' and 'key_changes_v1_0'.
However, the latter installs it without the 'IF NOT EXISTS'
condition, which make the query fail.

Solution: Change the queries in key_changes_v1_0 to have the
'IF NOT EXISTS' condition on the installation of indexes.
Note: even tho the only problematic one seems to be
'fki_rsc_category_id' I changed them all as it shouldn't create
any issue, but might prevent some in the future.