Make tinymce default to newest when the configured version is unsupported
Problem: It can happen that the configured version of the editor is
not available and no template can be included.
This can happen with misconfigurations but also with time: a version
that used to be available can become unsupported and updating the
mod_editor_tinymce would break the editor on such a website.

Solution: Make selecting the template to include more robust and
default to the newest version when the configured one can't be used.
Additionally, log a warning when this happens.