zotonic_mod_cookie_consent: carry nonce on script replacement
Problem: once cookies are accepted, the cookie_consent mod replaces
script tags that required consent with a new tag that gets executed.
However, if there was a nonce set up for the CSP headers this isn't
carried over, which results in a failure to run the script with the
default 'script-src' CSP directive.

Solution: modify the replacing code to add the nonce.