Exclude 'pivot_geocode_qhash' from rsc raw gets
Problem: 'pivot_geocode_qhash' can be any binary, including one that
is not valid UTF8, and attempting to export the resource/property
(via 'controller_api', for example) causes an error in the encoders,
which attempt to interpret it as a binary string.

Solution: As the for most other `pivot` properties, this can be
excluded from the rsc get (it's still accessible as a DB column).